move forms from view to forms file
