+++ main/forms.py
@@ -5,6 +5,7 @@ from main.models import UserProfile, MetaData
 from registration.models import RegistrationProfile
 from utils.country_field import COUNTRIES
 from django.forms import ModelForm
+from odk_viewer.models import DataDictionary
 import re
 
 FORM_LICENSES_CHOICES = (
@@ -120,3 +121,29 @@ class RegistrationFormUserProfile(RegistrationFormUniqueEmail, UserProfileFormRe
         UserProfileFormRegister.save(self, new_user)
         return new_user
 
+class SupportDocForm(forms.Form):
+    doc = forms.FileField(label="Supporting document", required=True)
+
+class QuickConverterFile(forms.Form):
+    xls_file = forms.FileField(label="XLS File", required=False)
+
+class QuickConverterURL(forms.Form):
+    xls_url = forms.URLField(verify_exists=False, label="XLS URL", required=False)
+
+class QuickConverter(QuickConverterFile, QuickConverterURL):
+    def publish(self, user):
+        if self.is_valid():
+            cleaned_xls_file = self.cleaned_data['xls_file']
+            if not cleaned_xls_file:
+                cleaned_url = self.cleaned_data['xls_url']
+                cleaned_xls_file = urlparse(cleaned_url)
+                cleaned_xls_file = '_'.join(cleaned_xls_file.path.split('/')[-2:])
+                if cleaned_xls_file[-4:] != '.xls':
+                    cleaned_xls_file += '.xls'
+                cleaned_xls_file = upload_to(None, cleaned_xls_file, user.username)
+                xls_data = ContentFile(urllib2.urlopen(cleaned_url).read())
+                cleaned_xls_file = default_storage.save(cleaned_xls_file, xls_data)
+            return DataDictionary.objects.create(
+                user=user,
+                xls=cleaned_xls_file
+                )
